requirements.txt
websocket-client==1.8.0
websockets==12.0
